feat: refactor error handling structure and update imports
