+++ src/errors/index.ts
